refactor(ir): extract the allocator bump walk into a shared SpaceFootprint primitive (#1475)
Move the ordered alignment bump walk (SpaceFootprint) and the system.reserve_buffer resolution (ResolveReserveBufferBases) into include/pypto/ir/transforms/utils/, and drive AllocateMemoryAddr's address emission through them. Behavior-preserving — the realized per-space footprint is unchanged — but the walk is now reusable so a downstream capacity check can be parity-by-construction with the allocator.
