+++ tests/ut/ir/transforms/test_allocate_memory_addr_pass.py
@@ -165,7 +165,10 @@ def test_allocate_memory_addr_rejects_overlapping_reserve_buffer_ranges():
             _overlap_buf = pl.reserve_buffer(name="overlap_slot_buffer", size=1024, base=2048)
 
     with pytest.raises(
-        Exception, match=re.escape("AllocateMemoryAddr found overlapping reserve_buffer ranges")
+        # Message is now emitted by the shared reserve_buffer_utils resolver (used by both
+        # AllocateMemoryAddr and MemoryReuse), so match the pass-agnostic substring.
+        Exception,
+        match=re.escape("overlapping reserve_buffer ranges"),
     ):
         program = passes.init_mem_ref()(Before)
         passes.allocate_memory_addr()(program)
